Apply fixes from analysis log

+++ lib/services/bet_slip_service.dart
@@ -57,7 +57,7 @@ class BetSlipService {
       id: ticketId,
       userId: userId,
       tips: tips,
-      stake: stake,
+      stake: stake.toDouble(),
       totalOdd: totalOdd,
       potentialWin: potentialWin,
       createdAt: DateTime.now(),
